Rename tokens to highlighter
