+++ app/src/lib/highlighter.ts
